fix: fix error with using array index as key. Page components folders reorganized.
