+++ src/pages/favorites-empty/favoritesEmptyPage.tsx
