+++ src/pages/favorites/favoritesPage.tsx
